+++ app/src/main/java/com/app/arcabyolimpo/data/repository/supplies/SupplyRepositoryImpl.kt
@@ -216,4 +216,51 @@ class SupplyRepositoryImpl
             } catch (e: Exception) {
                 Result.failure(e)
             }
+
+        override suspend fun updateSupply(
+            idSupply: String,
+            supply: SupplyAdd,
+            image: Uri?
+        ): Result<Unit> =
+            try {
+                val imagePart =
+                    image?.let {
+                        val input = context.contentResolver.openInputStream(it)
+                        val bytes = input?.readBytes()
+                        input?.close()
+                        if (bytes != null) {
+                            val requestFile =
+                                bytes.toRequestBody(
+                                    context.contentResolver
+                                        .getType(it)?.toMediaTypeOrNull(),
+                                )
+                            MultipartBody.Part.createFormData(
+                                "imagenInsumo",
+                                "image.jpg",
+                                requestFile,
+                            )
+                        } else {
+                            null
+                        }
+                    }
+
+                val idWorkshop = supply.idWorkshop.toRequestBody("text/plain".toMediaTypeOrNull())
+                val name = supply.name.toRequestBody("text/plain".toMediaTypeOrNull())
+                val measureUnit = supply.measureUnit.toRequestBody("text/plain".toMediaTypeOrNull())
+                val idCategory = supply.idCategory.toRequestBody("text/plain".toMediaTypeOrNull())
+                val status = supply.status.toString().toRequestBody("text/plain".toMediaTypeOrNull())
+
+                api.updateSupply(
+                    idSupply = idSupply,
+                    idWorkshop = idWorkshop,
+                    name = name,
+                    measureUnit = measureUnit,
+                    idCategory = idCategory,
+                    status = status,
+                    imagenInsumo = imagePart,
+                )
+                Result.success(Unit)
+            } catch (e: Exception) {
+                Result.failure(e)
+            }
     }