+++ CHANGELOG.md
@@ -5,6 +5,7 @@
    receive the URI value from the view model  
    
 ### Added
+- US for modify Supply with MVVM and clean architecture principles
 - Delete supply batch
 - Added filter and order products