+++ CHANGELOG.md
@@ -1,10 +1,28 @@
+## [0.1.0] - 17/11/2025
+
+### Added
+- Delete supply batch
+- Added filter and order products
+- Added filter and order product batches
+- Added functional product batch search bar
+
 ## [0.1.0] - 16/11/2025
 
 ### Added
- - Added products search
- - Add DeleteWorkshop Logic, Data and Domain
- - Add way to get users for workshops
- - Added functionality search beneficiaries
+- Added products search
+- Add DeleteWorkshop Logic, Data and Domain
+- Add way to get users for workshops
+- Added functionality search beneficiaries
+- Added functionality view a Workshop (Video functionality needs more info)
+- Filter and order beneficiaries functionality
+- Fixed merge conflicts
+- Added US to Delete Product Batches
+
+### Fixed
+- Refreshing on Product Batches List
+
+## [0.1.0] - 15/11/2025
+### Fixed
 
 ## [0.1.0] - 15/11/2025